feat: add `BREAK` statement support
- Add support for `BREAK` statement inside `WHILE` loops.
- `BREAK` statements must be inside a `WHILE` loop, otherwise it will throw an error.
- `BREAK` statements will jump to the end of the `WHILE` loop.

The `BREAK` statement can now be used to exit a `WHILE` loop prematurely. This is useful for situations where you want to stop the loop based on a specific condition. For example:

```
while true do
  if something something_else = do
    break
  end
end
```

In this example, the loop will continue to run as long as `something` is not equal to `something_else`. However, if the condition is met, the `BREAK` statement will be executed, and the loop will be exited.

The `BREAK` statement can only be used inside a `WHILE` loop. If you try to use it outside of a `WHILE` loop, an error will be thrown.

This commit also includes a fix for a bug where the `jump_ip` attribute of the `BREAK` token was not being set correctly. This could lead to the loop not being exited correctly.