feat: Add token type for set
- Adds a new token type `SET` to the `TokenType` enum.
- Updates the `Token` constructor to handle the new token type.

This commit introduces a new token type `SET` which allows to track changes for a set of keys. This is important when dealing with sets of keys and is needed to keep track of the keys that have been added, removed or updated.

The `SET` token type is added to the `TokenType` enum. The `Token` constructor and class are updated to handle the new token type.